Change semantics of security groups creation
This is only a refactor, and does not change the way it actually works.

I changed it to check for `application` to make it more explicit that it
only gets created when we're using this module as an ALB.

This will also make it easier to understand when I start adding more
logic to the module later.